[Patch v6.9.36] Split main helpers
