+++ AGENTS.md
@@ -347,3 +347,4 @@
 - Added `wfv_aggregator` module for fold result aggregation.
 - Added `state_manager` module for persistent system state management.
 - Added `main_helpers`, `model_helpers`, and `pipeline_helpers` modules to organize main functions.
+- Added `main_utils` module for CSV loading and output directory helpers.
