+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 - New/Updated unit tests added for src/main_helpers.py and others
 - QA: pytest -q passed (tests count TBD)
 
+### 2025-06-15
+- [Patch v6.9.36] Extract helper functions to new modules
+- New/Updated unit tests added for N/A
+- QA: pytest -q passed (tests count TBD)
+
 # ### 2025-07-05
 - [Patch v6.9.33] ปรับปรุงตัวแปลงวันที่ใน auto_convert_gold_csv
 - New/Updated unit tests added for tests/test_auto_convert_csv.py
